+++ src/logic/shipLogic.ts
@@ -157,64 +157,6 @@ export const checkShieldDomeLimit = (
 }
 
 /**
- * 计算导弹发射井容量
- */
-export const calculateMissileSiloCapacity = (buildings: Partial<Record<BuildingType, number>>): number => {
-  const siloLevel = buildings[BuildingType.MissileSilo] || 0
-  return siloLevel * 10 // 每级存储10枚导弹
-}
-
-/**
- * 计算当前导弹总数
- */
-export const calculateCurrentMissileCount = (defense: Partial<Record<DefenseType, number>>): number => {
-  const interplanetaryMissiles = defense[DefenseType.InterplanetaryMissile] || 0
-  const antiBallisticMissiles = defense[DefenseType.AntiBallisticMissile] || 0
-  return interplanetaryMissiles + antiBallisticMissiles
-}
-
-/**
- * 计算建造队列中的导弹总数
- */
-export const calculateQueueMissileCount = (buildQueue: Array<{ type: string; itemType: string; quantity?: number }>): number => {
-  let queueMissileCount = 0
-
-  for (const item of buildQueue) {
-    if (item.type === 'defense') {
-      const defenseType = item.itemType as DefenseType
-      if (defenseType === DefenseType.InterplanetaryMissile || defenseType === DefenseType.AntiBallisticMissile) {
-        queueMissileCount += item.quantity || 0
-      }
-    }
-  }
-
-  return queueMissileCount
-}
-
-/**
- * 检查导弹容量限制
- */
-export const checkMissileSiloLimit = (
-  defenseType: DefenseType,
-  currentDefense: Partial<Record<DefenseType, number>>,
-  buildings: Partial<Record<BuildingType, number>>,
-  quantity: number,
-  buildQueue?: Array<{ type: string; itemType: string; quantity?: number }>
-): boolean => {
-  // 只对导弹类型进行检查
-  if (defenseType !== DefenseType.InterplanetaryMissile && defenseType !== DefenseType.AntiBallisticMissile) {
-    return true
-  }
-
-  const maxCapacity = calculateMissileSiloCapacity(buildings)
-  const currentCount = calculateCurrentMissileCount(currentDefense)
-  const queueCount = buildQueue ? calculateQueueMissileCount(buildQueue) : 0
-  const newCount = currentCount + queueCount + quantity
-
-  return newCount <= maxCapacity
-}
-
-/**
  * 创建舰船建造队列项
  */
 export const createShipQueueItem = (shipType: ShipType, quantity: number, buildTime: number): BuildQueueItem => {
